@@ -1,22 +1,31 @@
-from django.conf.urls import include, url
+from django.conf.urls import include
+from django.urls import re_path
 from django.views.generic import TemplateView
 from wagtail.admin import urls as wagtailadmin_urls
 from wagtail.core import urls as wagtail_urls
 from wagtailmenus.api import urls as api_urls
 
 urlpatterns = [
-    url(r'^admin/', include(wagtailadmin_urls)),
-    url(r'^custom-url/$', TemplateView.as_view(template_name='page.html')),
-    url(r'^sub_menu-tag-used-directly/$',
-        TemplateView.as_view(template_name='sub_menu-tag-used-directly.html')),
-    url(r'^superheroes/marvel-comics/custom-man/about/$',
-        TemplateView.as_view(template_name='page.html')),
-    url(r'^about-us/meet-the-team/staff-member-one/$',
-        TemplateView.as_view(template_name='page.html')),
-    url(r'^about-us/meet-the-team/custom-url/$',
-        TemplateView.as_view(template_name='page.html')),
-    url(r'^news-and-events/$',
-        TemplateView.as_view(template_name='page.html')),
-    url(r'menu_api/', include('wagtailmenus.api.urls')),
-    url(r'', include(wagtail_urls)),
+    re_path(r"^admin/", include(wagtailadmin_urls)),
+    re_path(r"^custom-url/$", TemplateView.as_view(template_name="page.html")),
+    re_path(
+        r"^sub_menu-tag-used-directly/$",
+        TemplateView.as_view(template_name="sub_menu-tag-used-directly.html"),
+    ),
+    re_path(
+        r"^superheroes/marvel-comics/custom-man/about/$",
+        TemplateView.as_view(template_name="page.html"),
+    ),
+    re_path(
+        r"^about-us/meet-the-team/staff-member-one/$",
+        TemplateView.as_view(template_name="page.html"),
+    ),
+    re_path(
+        r"^about-us/meet-the-team/custom-url/$",
+        TemplateView.as_view(template_name="page.html"),
+    ),
+    re_path(r"^news-and-events/$", TemplateView.as_view(template_name="page.html")),
+    url(r"menu_api/", include("wagtailmenus.api.urls")),
+    url(r"", include(wagtail_urls)),
+    re_path(r"", include(wagtail_urls)),
 ]
