@@ -5,15 +5,21 @@ from django.contrib.admin.utils import quote, unquote
 from django.core.exceptions import PermissionDenied
 from django.shortcuts import get_object_or_404, redirect
 from django.utils.text import capfirst
-from django.utils.translation import ugettext as _
+from django.utils.translation import gettext as _
 from wagtail.admin import messages
-from wagtail.admin.edit_handlers import ObjectList, TabbedInterface
 from wagtail.contrib.modeladmin.views import (
     WMABaseView, CreateView, EditView, ModelFormView
 )
-from wagtail.core.models import Site
-
 from wagtailmenus.conf import settings
+from distutils.version import LooseVersion
+try:
+    from wagtail import __version__ as wagtail_version
+    from wagtail.models import Site
+    from wagtail.admin.panels import ObjectList, TabbedInterface
+except ImportError:
+    from wagtail.core import __version__ as wagtail_version
+    from wagtail.core.models import Site
+    from wagtail.admin.edit_handlers import ObjectList, TabbedInterface
 
 
 class SiteSwitchForm(forms.Form):
@@ -54,8 +60,8 @@ class MenuTabbedInterfaceMixin:
                 ObjectList(self.model.settings_panels, heading=_("Settings"),
                            classname="settings"),
             ])
-        if hasattr(edit_handler, 'bind_to'):
-            # For Wagtail>=2.5
+        if LooseVersion(wagtail_version) < LooseVersion('3.0'):
+            # For Wagtail>=2.5,<3.0
             return edit_handler.bind_to(model=self.model)
         return edit_handler.bind_to_model(self.model)
 
