@@ -1,15 +1,28 @@
 from django.db import models
 from modelcluster.fields import ParentalKey
 from rest_framework.fields import CharField
-from wagtail.admin.edit_handlers import (
-    FieldPanel, MultiFieldPanel, PageChooserPanel)
+from wagtail.admin.edit_handlers import FieldPanel, MultiFieldPanel, PageChooserPanel
 from wagtail.api import APIField
 
 from wagtailmenus.models import (
-    SectionMenu, ChildrenMenu, AbstractMainMenu,
-    AbstractMainMenuItem, AbstractFlatMenu, AbstractFlatMenuItem)
+    SectionMenu,
+    ChildrenMenu,
+    AbstractMainMenu,
+    AbstractMainMenuItem,
+    AbstractFlatMenu,
+    AbstractFlatMenuItem,
+)
 
 from .utils import TranslatedField
+
+try:
+    from wagtail.admin.panels import FieldPanel, MultiFieldPanel, PageChooserPanel
+except ImportError:
+    from wagtail.admin.edit_handlers import (
+        FieldPanel,
+        MultiFieldPanel,
+        PageChooserPanel,
+    )
 
 
 class CustomChildrenMenu(ChildrenMenu):
@@ -22,38 +35,38 @@ class CustomSectionMenu(SectionMenu):
 
 class MultilingualMenuItem(models.Model):
     link_text_de = models.CharField(
-        verbose_name='link text (de)',
+        verbose_name="link text (de)",
         max_length=255,
         blank=True,
     )
     link_text_fr = models.CharField(
-        verbose_name='link text (fr)',
+        verbose_name="link text (fr)",
         max_length=255,
         blank=True,
     )
-    translated_link_text = TranslatedField(
-        'link_text', 'link_text_de', 'link_text_fr'
-    )
+    translated_link_text = TranslatedField("link_text", "link_text_de", "link_text_fr")
 
     class Meta:
         abstract = True
-        ordering = ('sort_order',)
+        ordering = ("sort_order",)
 
     @property
     def menu_text(self):
-        return self.translated_link_text or getattr(
-            self.link_page, 'translated_title', None
-        ) or self.link_page.title
+        return (
+            self.translated_link_text
+            or getattr(self.link_page, "translated_title", None)
+            or self.link_page.title
+        )
 
     panels = (
-        PageChooserPanel('link_page'),
-        FieldPanel('link_url'),
-        FieldPanel('url_append'),
-        FieldPanel('link_text'),
-        FieldPanel('link_text_de'),
-        FieldPanel('link_text_fr'),
-        FieldPanel('handle'),
-        FieldPanel('allow_subnav'),
+        PageChooserPanel("link_page"),
+        FieldPanel("link_url"),
+        FieldPanel("url_append"),
+        FieldPanel("link_text"),
+        FieldPanel("link_text_de"),
+        FieldPanel("link_text_fr"),
+        FieldPanel("handle"),
+        FieldPanel("allow_subnav"),
     )
 
 
@@ -63,10 +76,11 @@ class MainMenuCustomMenuItem(MultilingualMenuItem, AbstractMainMenuItem):
 
     `WAGTAILMENUS_MAIN_MENU_ITEMS_RELATED_NAME = 'custom_menu_items'
     """
+
     menu = ParentalKey(
-        'wagtailmenus.MainMenu',
+        "wagtailmenus.MainMenu",
         on_delete=models.CASCADE,
-        related_name="custom_menu_items"
+        related_name="custom_menu_items",
     )
 
 
@@ -78,9 +92,9 @@ class FlatMenuCustomMenuItem(MultilingualMenuItem, AbstractFlatMenuItem):
     """
 
     menu = ParentalKey(
-        'wagtailmenus.FlatMenu',
+        "wagtailmenus.FlatMenu",
         on_delete=models.CASCADE,
-        related_name="custom_menu_items"
+        related_name="custom_menu_items",
     )
 
 
@@ -90,43 +104,41 @@ class CustomMainMenu(AbstractMainMenu):
 
 class CustomFlatMenu(AbstractFlatMenu):
     heading_de = models.CharField(
-        verbose_name='heading (de)',
+        verbose_name="heading (de)",
         max_length=255,
         blank=True,
     )
     heading_fr = models.CharField(
-        verbose_name='heading (fr)',
+        verbose_name="heading (fr)",
         max_length=255,
         blank=True,
     )
-    translated_heading = TranslatedField(
-        'heading', 'heading_de', 'heading_fr'
-    )
+    translated_heading = TranslatedField("heading", "heading_de", "heading_fr")
 
     content_panels = (
         MultiFieldPanel(
             heading="Settings",
             children=(
-                FieldPanel('title'),
-                FieldPanel('site'),
-                FieldPanel('handle'),
-            )
+                FieldPanel("title"),
+                FieldPanel("site"),
+                FieldPanel("handle"),
+            ),
         ),
         MultiFieldPanel(
             heading="Heading",
             children=(
-                FieldPanel('heading'),
-                FieldPanel('heading_de'),
-                FieldPanel('heading_fr'),
+                FieldPanel("heading"),
+                FieldPanel("heading_de"),
+                FieldPanel("heading_fr"),
             ),
-            classname='collapsible'
+            classname="collapsible",
         ),
         AbstractFlatMenu.content_panels[1],
     )
 
     api_fields = [
-        APIField('handle'),
-        APIField('translated_heading', CharField(read_only=True)),
+        APIField("handle"),
+        APIField("translated_heading", CharField(read_only=True)),
     ]
 
 
@@ -138,9 +150,7 @@ class CustomMainMenuItem(MultilingualMenuItem, AbstractMainMenuItem):
     overridden ('menu_items' is the default value)."""
 
     menu = ParentalKey(
-        'CustomMainMenu',
-        on_delete=models.CASCADE,
-        related_name="menu_items"
+        "CustomMainMenu", on_delete=models.CASCADE, related_name="menu_items"
     )
 
 
@@ -152,7 +162,5 @@ class CustomFlatMenuItem(MultilingualMenuItem, AbstractFlatMenuItem):
     overridden ('menu_items' is the default value)."""
 
     menu = ParentalKey(
-        'CustomFlatMenu',
-        on_delete=models.CASCADE,
-        related_name="menu_items"
+        "CustomFlatMenu", on_delete=models.CASCADE, related_name="menu_items"
     )
