@@ -3,14 +3,19 @@ from datetime import date
 from django.db import models
 from django.http import Http404
 from django.template.response import TemplateResponse
-from wagtail.admin.edit_handlers import (
-    FieldPanel, MultiFieldPanel, PublishingPanel
-)
 from wagtail.contrib.routable_page.models import RoutablePageMixin, route
-from wagtail.core.models import Page
-
 from wagtailmenus.models import MenuPage, AbstractLinkPage
 from .utils import TranslatedField
+try:
+    from wagtail.admin.panels import (
+        FieldPanel, MultiFieldPanel, PublishingPanel
+    )
+    from wagtail.models import Page
+except ImportError:
+    from wagtail.admin.edit_handlers import (
+        FieldPanel, MultiFieldPanel, PublishingPanel
+    )
+    from wagtail.core.models import Page
 
 
 class MultilingualMenuPage(MenuPage):
