@@ -1,15 +1,20 @@
+from collections import OrderedDict
+
 from django.contrib.auth import get_user_model
 from django.contrib.auth.models import Group
 from django.contrib.contenttypes.models import ContentType
 from django.test import TestCase, TransactionTestCase, override_settings
 
 from django_webtest import WebTest
-from wagtail.admin.edit_handlers import ObjectList
-from wagtail.core.models import Page, Site
-
 from wagtailmenus import get_flat_menu_model, get_main_menu_model
 
 from wagtailmenus.tests.models import LinkPage
+try:
+    from wagtail.admin.panels import ObjectList
+    from wagtail.models import Page, Site
+except ImportError:
+    from wagtail.admin.edit_handlers import ObjectList
+    from wagtail.core.models import Page, Site
 
 FlatMenu = get_flat_menu_model()
 
@@ -40,7 +45,7 @@ class CMSUsecaseTests(WebTest):
             '%scopy/%s/' % (self.base_flatmenu_admin_url, site_one_footer_menu.pk),
             user='test1')
 
-        form = copy_view.forms[1]
+        form = copy_view.forms[next(reversed(OrderedDict(copy_view.forms)))]
         form['site'] = site_two.pk
         response = form.submit().follow()
 
@@ -71,7 +76,7 @@ class CMSUsecaseTests(WebTest):
         copy_view = self.app.get(
             '%scopy/%s/' % (self.base_flatmenu_admin_url, site_one_footer_menu.pk),
             user='test1')
-        form = copy_view.forms[1]
+        form = copy_view.forms[next(reversed(OrderedDict(copy_view.forms)))]
         form['site'] = site_two.pk
         response = form.submit()
 
@@ -81,7 +86,7 @@ class CMSUsecaseTests(WebTest):
     def test_main_menu_save_success(self):
         edit_view = self.app.get(
             '%sedit/1/' % self.base_mainmenu_admin_url, user='test1')
-        form = edit_view.forms[2]
+        form = edit_view.forms[next(reversed(OrderedDict(edit_view.forms)))]
         response = form.submit().follow()
 
         assert 'Main menu updated successfully.' in response
@@ -259,5 +264,9 @@ class TestNonSuperUser(TransactionTestCase):
         self.client.login(username='test2', password='password')
 
     def test_mainmenu_edit_denied(self):
-        response = self.client.get('/admin/wagtailmenus/mainmenu/edit/1/')
-        self.assertEqual(response.status_code, 403)
+        response = self.client.get(
+            "/admin/wagtailmenus/mainmenu/edit/1/", follow=True
+        )
+        self.assertContains(
+            response, "Sorry, you do not have permission to access this area"
+        )
