@@ -1,9 +1,10 @@
 from django.test import TestCase
-
-from wagtail.core.models import Page, Site
-
 from wagtailmenus.errors import RequestUnavailableError
 from wagtailmenus.models import Menu, MenuWithMenuItems
+try:
+    from wagtail.models import Page, Site
+except ImportError:
+    from wagtail.core.models import Page, Site
 
 
 class TestMenuGetSite(TestCase):
