@@ -1,7 +1,10 @@
 from django.test import TestCase
 from django.core.management import call_command
-from wagtail.core.models import Site
 from wagtailmenus.conf import settings
+try:
+    from wagtail.models import Site
+except ImportError:
+    from wagtail.core.models import Site
 
 
 class TestAutoPopulateMainMenus(TestCase):
