@@ -2,8 +2,10 @@ from __future__ import absolute_import, unicode_literals
 
 from bs4 import BeautifulSoup
 from django.test import TestCase
-from wagtail.core import hooks
-
+try:
+    from wagtail import hooks
+except ImportError:
+    from wagtail.core import hooks
 
 class TestHooks(TestCase):
     fixtures = ['test.json']
