@@ -1,12 +1,14 @@
 from django.core.exceptions import ValidationError
 from django.test import TestCase
 from django.test.client import RequestFactory
-from wagtail.core.models import Page
-
 from wagtailmenus.conf import settings
 from wagtailmenus.models import (
     AbstractMenuItem, MainMenu, MainMenuItem, FlatMenu, FlatMenuItem
 )
+try:
+    from wagtail.models import Page
+except ImportError:
+    from wagtail.core.models import Page
 
 
 class MenuItemModelTestMixin:
@@ -105,7 +107,7 @@ class TestMenuItemsForRequest(TestCase):
         self.rf = RequestFactory()
 
     def test_menu_item_exact_match(self):
-        menu_item = AbstractMenuItem()
+        menu_item = MainMenuItem()
         menu_item.link_url = '/some-page/'
         request = self.rf.get('/some-page/')
 
@@ -113,7 +115,7 @@ class TestMenuItemsForRequest(TestCase):
         self.assertEqual(active_class, settings.ACTIVE_CLASS)
 
     def test_menu_item_with_query_params(self):
-        menu_item = AbstractMenuItem()
+        menu_item = MainMenuItem()
         menu_item.link_url = '/some-page/?some_param=foo'
         request = self.rf.get('/some-page/?some_other_param=foo')
 
@@ -121,7 +123,7 @@ class TestMenuItemsForRequest(TestCase):
         self.assertEqual(active_class, settings.ACTIVE_CLASS)
 
     def test_menu_item_with_hash_fragment(self):
-        menu_item = AbstractMenuItem()
+        menu_item = MainMenuItem()
         menu_item.link_url = '/some-page/#some_fragment'
         request = self.rf.get('/some-page/#some_other_hash_fragment')
 
@@ -129,7 +131,7 @@ class TestMenuItemsForRequest(TestCase):
         self.assertEqual(active_class, settings.ACTIVE_CLASS)
 
     def test_menu_item_ancestor(self):
-        menu_item = AbstractMenuItem()
+        menu_item = MainMenuItem()
         menu_item.link_url = '/some-page/'
         request = self.rf.get('/some-page/some-child-page')
 
@@ -137,7 +139,7 @@ class TestMenuItemsForRequest(TestCase):
         self.assertEqual(active_class, settings.ACTIVE_ANCESTOR_CLASS)
 
     def test_menu_item_with_netloc(self):
-        menu_item = AbstractMenuItem()
+        menu_item = MainMenuItem()
         menu_item.link_url = 'https://example.com/some-page'
         request = self.rf.get('/some-page/')
 
