@@ -1,10 +1,12 @@
 from bs4 import BeautifulSoup
 from django.test import TestCase, override_settings
-from wagtail.core.models import Site
-
 from wagtailmenus.errors import SubMenuUsageError
 from wagtailmenus.models import MainMenu, FlatMenu
 from wagtailmenus.templatetags.menu_tags import validate_supplied_values
+try:
+    from wagtail.models import Site
+except ImportError:
+    from wagtail.core.models import Site
 
 
 class TestTemplateTags(TestCase):
