@@ -1,8 +1,11 @@
 from django.http import Http404, HttpRequest
-from wagtail.core.models import Page, Site
 
 from wagtailmenus.conf import settings
 from wagtailmenus.models.menuitems import MenuItem
+try:
+    from wagtail.models import Page, Site
+except ImportError:
+    from wagtail.core.models import Page, Site
 
 
 def get_fake_site():
