@@ -1,4 +1,3 @@
-from wagtail.core import hooks
 from wagtail.contrib.modeladmin.options import modeladmin_register
 
 from wagtailmenus.conf import settings
@@ -6,6 +5,10 @@ from wagtailmenus.utils.misc import derive_section_root
 from wagtailmenus.modeladmin import ( # noqa
     MainMenuAdmin, FlatMenuAdmin, FlatMenuButtonHelper
 )
+try:
+    from wagtail import hooks
+except ImportError:
+    from wagtail.core import hooks
 
 if settings.MAIN_MENUS_EDITABLE_IN_WAGTAILADMIN:
     modeladmin_register(settings.objects.MAIN_MENUS_MODELADMIN_CLASS)
